Add link to elm community resources
I'd prefer the text to only point people to the elm community page in general, and not specifically mention the mailing list. However, I decided not to delete reference to the mailing page because simply adding a reference to the elm community page seems more likely to get accepted and merged.